Refs #1648. Extends signal monitoring flags from 32 bits to 64 bits so that more things can be monitored.
This is needed to add support for STB power monitoring for Firewire set top boxes.

git-svn-id: http://svn.mythtv.org/svn/trunk@12222 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/dtvsignalmonitor.cpp
@@ -119,13 +119,13 @@ QStringList DTVSignalMonitor::GetStatusList(bool kick)
     return list;
 }
 
-void DTVSignalMonitor::AddFlags(uint _flags)
+void DTVSignalMonitor::AddFlags(uint64_t _flags)
 {
     SignalMonitor::AddFlags(_flags);
     UpdateMonitorValues();
 }
 
-void DTVSignalMonitor::RemoveFlags(uint _flags)
+void DTVSignalMonitor::RemoveFlags(uint64_t _flags)
 {
     SignalMonitor::RemoveFlags(_flags);
     UpdateMonitorValues();
